% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flusight_bin_endpoints.R
\name{get_flusight_bin_endpoints}
\alias{get_flusight_bin_endpoints}
\title{Get the bin endpoints for FluSight categorical targets}
\usage{
get_flusight_bin_endpoints(target_ts, location_meta)
}
\arguments{
\item{target_ts}{Data frame with target time series data for FluSight,
including the columns \code{location}, \code{date}, and \code{value}}

\item{location_meta}{Data frame with metadata about locations for FluSight,
matching the format of the file in cdcepi/FluSight-forecast-hub/auxiliary-data/locations.csv}
}
\value{
a data frame
}
\description{
Get the bin endpoints for FluSight categorical targets
}
