% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idforecastutils-package.R
\docType{package}
\name{idforecastutils-package}
\alias{idforecastutils}
\alias{idforecastutils-package}
\title{idforecastutils: Utility functions for general infectious disease forecasting use}
\description{
A set of utility functions for various infectious disease forecasting models, exercises, and rounds
}
\author{
\strong{Maintainer}: Li Shandross \email{lshandross@umass.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
