% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_format_daily_data.R
\name{load_format_daily_data}
\alias{load_format_daily_data}
\title{Load and format infectious disease target data queried from covidData}
\usage{
load_format_daily_data(disease, locations_df = NULL, due_date = NULL)
}
\arguments{
\item{disease}{a string of the diseased query data for. Must be one of "flu"
(flu hospitalizations) or "covid" (covid hospitalizations).}

\item{locations_df}{a data frame of locations to forecast for. Must contain
the following columns: \code{geo_value} (lowercase abbreviations), and \code{location}
(fips codes). Defaults to NULL.}

\item{due_date}{a string specifying the date as of for which to query the data}
}
\value{
a data frame of daily data for the specified disease with the
following columns: \code{geo_value}, \code{location}, \code{time_value}, \code{value}
}
\description{
Load and format infectious disease target data queried from covidData
}
